+++ plugins/alignments/src/MismatchParser/index.test.ts
@@ -5,7 +5,7 @@ import {
   parseCigar,
   getNextRefPos,
   getModificationPositions,
-} from './MismatchParser'
+} from './index'
 
 const seq =
   'AAAAAAAAAACAAAAAAAAAAAAAACCCCCCCCCCCCCCCCCCCCCCCCCGGGGGGGGGGGGGGGGGGGGGGGGGTTTTTTTTTTTTTTTTTTTTTTTTT'
