+++ plugins/linear-comparative-view/src/LinearReadVsRef/util.ts
@@ -1,58 +1,3 @@
-import { Feature } from '@jbrowse/core/util'
-import { MismatchParser } from '@jbrowse/plugin-alignments'
-
-const { parseCigar } = MismatchParser
-
-export function getLengthOnRef(cigar: string) {
-  const cigarOps = parseCigar(cigar)
-  let lengthOnRef = 0
-  for (let i = 0; i < cigarOps.length; i += 2) {
-    const len = +cigarOps[i]
-    const op = cigarOps[i + 1]
-    if (op !== 'H' && op !== 'S' && op !== 'I') {
-      lengthOnRef += len
-    }
-  }
-  return lengthOnRef
-}
-
-export function getLength(cigar: string) {
-  const cigarOps = parseCigar(cigar)
-  let length = 0
-  for (let i = 0; i < cigarOps.length; i += 2) {
-    const len = +cigarOps[i]
-    const op = cigarOps[i + 1]
-    if (op !== 'D' && op !== 'N') {
-      length += len
-    }
-  }
-  return length
-}
-
-export function getLengthSansClipping(cigar: string) {
-  const cigarOps = parseCigar(cigar)
-  let length = 0
-  for (let i = 0; i < cigarOps.length; i += 2) {
-    const len = +cigarOps[i]
-    const op = cigarOps[i + 1]
-    if (op !== 'H' && op !== 'S' && op !== 'D' && op !== 'N') {
-      length += len
-    }
-  }
-  return length
-}
-
-export function getClip(cigar: string, strand: number) {
-  return strand === -1
-    ? +(cigar.match(/(\d+)[SH]$/) || [])[1] || 0
-    : +(cigar.match(/^(\d+)([SH])/) || [])[1] || 0
-}
-
-export function getTag(f: Feature, tag: string) {
-  const tags = f.get('tags')
-  return tags ? tags[tag] : f.get(tag)
-}
-
 export function mergeIntervals<T extends { start: number; end: number }>(
   intervals: T[],
   w = 5000,
